Moved definition of LinkClock into hpp, use static methods for init
